Added handling of /var/lib/sss/pipes/nss file
By default, if the file exists, exception is thrown.
You can set -Djigawatts.ssspipesnss to ignore to behave as before - to
die cryptically
Also you may set it to force, which will try to remove file before criu
actions and resore afterwards. If anything goes wrong, it throws.
Last choice is fierce, which is as force, but do not die if removal or
restore of file goes bad